requirements.txt
apt-xapian-index==0.49
asgiref==3.5.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-crispy-forms==1.14.0
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.3
galternatives==1.0.6
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lubuntu-update-notifier==0.1
MarkupSafe==2.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pexpect==4.6.0
Pillow==7.0.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.8
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.2.4
systemd-python==234
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
whitenoise==6.0.0
xkit==0.0.0
